🐛 fix(parser): update version parsing to support decimal values
- change version parsing logic from int to float
- allow for non-integer version numbers in metadata parsing

+++ Scripts/Common/WNCParser.cs
@@ -123,7 +123,7 @@ public static class WNCParser
   private static void ParseChartFormatLine(string line, ChartMetadata meta)
   {
     if (ParserUtils.TryParseProperty(line, "Version:", out string version))
-      meta.VERSION = int.TryParse(version, out int v) ? v : 0;
+      meta.VERSION = float.TryParse(version, out float v) ? v : 0;
   }
 
   // ── METADATA ──
